Add the main library code and update README file.
The library, despite being usable, is only able to correctly parse lyrics defined with absolute timings. Relative timing logic will be added in the future.
It also lacks information about the correct time to print and erase each sentence on the screen, meaning that at the moment it's only possible to know when the first syllable of each sentence is supposed to be sung.
The README file is updated with the purpose and current status of the project.